2230326-bacvkend-general Тестовый коммит 6.

+++ api/public/local/modules/rebit.share/lib/Infrastructure/Bitrix/Module/ModuleRoutingTrait.php
@@ -99,6 +99,10 @@ trait ModuleRoutingTrait
      */
     private function removeRouteSymlink(): void
     {
+        if (!is_link($this->routeFilePath)) {
+            return;
+        }
+
         if (!unlink($this->routeFilePath)) {
             Log::error("Не удалось удалить симлинк {$this->routeFilePath}");
         }
